Move read node types to separate file
